change version and add log
